♻️ refactor : Color 클래스 외부로 추출
