+++ src/main/java/softeer2nd/domain/chess/DefaultChessPointCalculator.java
@@ -6,7 +6,6 @@ import static softeer2nd.domain.chess.ChessGame.WIDTH;
 import java.util.List;
 import java.util.stream.IntStream;
 import softeer2nd.domain.chess.pieces.Piece;
-import softeer2nd.domain.chess.pieces.Piece.Color;
 
 public class DefaultChessPointCalculator implements ChessPointCalculator {
     private static final double DUPLICATE_FILE_PAWN_POINT = 0.5;
