+++ src/main/java/softeer2nd/domain/chess/pieces/Knight.java
@@ -2,6 +2,7 @@ package softeer2nd.domain.chess.pieces;
 
 import java.util.ArrayList;
 import java.util.List;
+import softeer2nd.domain.chess.Color;
 
 public class Knight extends Piece {
     private static final String REPRESENTATION = "n";
